+++ spec/models/my_inherited_doc.rb
@@ -1,0 +1,2 @@
+class MyInheritedDoc < MyDoc
+end